Changes to rankings displays and playoff result calculation for 2024 rules
