+++ src/components/RankingDisplay/TeamRankings/styles.module.scss
@@ -48,6 +48,10 @@
 
     td {
       padding-inline: 15px;
+
+      sup {
+        font-size: 0.6em;
+      }
     }
 
     tr:nth-child(even) {
